### 2019-04-14 #1377 @abatten
The ORCID number seems to be correct. It might have failed because my ORCID account didn't have my middle initial. I've changed it now so hopefully it works.
