### 2019-04-15 #1377 @arfon
I'm not sure why Whedon is struggling here. I've compiled the PDF locally: [10.21105.joss.01377.pdf](https://github.com/openjournals/joss-reviews/files/3081403/10.21105.joss.01377.pdf)
