### 2019-04-16 #1377 @abatten
Thanks @arfon. Looks like parts of the document didn't render as I'd expect them to. e.g missing citations, listing dot points and figure captions. I'll fix those up.
