### 2019-04-17 #1377 @arfon
I'm wondering if Whedon is struggling because of the size of the logo: https://github.com/abatten/fruitbat/blob/master/logo/fruitbat_logo.png
