### 2019-04-17 #1377 @abatten
It could be. The logo isn't super important to the paper, just thought it would be nice. I'll remove it and see.
