### 2019-04-17 #1377 @arfon
Anyway, here's an update proof: [10.21105.joss.01377.pdf](https://github.com/openjournals/joss-reviews/files/3087480/10.21105.joss.01377.pdf)
