### 2019-04-17 #1377 @arfon
> It could be. The logo isn't super important to the paper, just thought it would be nice. I'll remove it and see.

Or just resize it? If you want to try out the change, you can push a modification to a branch and then ask Whedon to compile the paper from that branch for testing (e.g. `@whedon generate pdf from branch testing`)
