### 2019-04-17 #1377 @abatten
It still seems to be struggling even without the logo. It could be the other figure, but it isn't that big. I'm not sure what the issue is.
