### 2019-04-17 31945aa
Fixed Planck13 citation and figure caption
