+++ paper/paper.bib
@@ -281,7 +281,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Planck2013,
+@ARTICLE{Planck13,
        author = {{Planck Collaboration} and {Ade}, P.~A.~R. and {Aghanim}, N. and
          {Armitage-Caplan}, C. and {Arnaud}, M. and {Ashdown}, M. and {Atrio-Barand
         ela}, F. and {Aumont}, J. and {Baccigalupi}, C. and {Banday}, A.~J. and
