### 2019-04-18 #1377 @arfon
It looks like the very long author lists on the Planck papers is blowing up the memory usage on the Heroku PDF worker:

```
2019-04-18T11:20:05.697913+00:00 heroku[router]: at=info method=POST path="/dispatch" host=whedon.herokuapp.com request_id=c5739468-eeda-43e3-bf27-38b7cab4f6dc fwd="140.82.115.12" dyno=web.1 connect=0ms service=2747ms status=200 bytes=253 protocol=https
2019-04-18T11:20:05.711888+00:00 app[worker.1]: 4 TID-1bkjao PDFWorker JID-21ab8dd21ac0cd55cb66681a INFO: start
2019-04-18T11:20:07.406918+00:00 heroku[router]: at=info method=POST path="/dispatch" host=whedon.herokuapp.com request_id=e499a364-b79b-4670-89e4-cc8f7d9e07e2 fwd="140.82.115.5" dyno=web.1 connect=0ms service=2002ms status=200 bytes=228 protocol=https
2019-04-18T11:20:48.119861+00:00 heroku[worker.1]: Process running mem=737M(142.2%)
2019-04-18T11:20:48.121062+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:21:06.931745+00:00 heroku[worker.1]: Process running mem=728M(142.2%)
2019-04-18T11:21:06.931745+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:21:28.033900+00:00 heroku[worker.1]: Process running mem=728M(142.2%)
2019-04-18T11:21:28.033981+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:21:46.996705+00:00 heroku[worker.1]: Process running mem=853M(166.6%)
2019-04-18T11:21:46.996784+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:22:07.247549+00:00 heroku[worker.1]: Process running mem=861M(168.2%)
2019-04-18T11:22:07.247659+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:22:27.579594+00:00 heroku[worker.1]: Process running mem=1020M(199.3%)
2019-04-18T11:22:27.579731+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:22:47.066625+00:00 heroku[worker.1]: Process running mem=1020M(199.3%)
2019-04-18T11:22:47.066721+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:23:06.512981+00:00 heroku[worker.1]: Process running mem=1021M(199.3%)
2019-04-18T11:23:06.513116+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:23:26.883818+00:00 heroku[worker.1]: Process running mem=1020M(199.3%)
2019-04-18T11:23:26.883865+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:23:46.561362+00:00 heroku[worker.1]: Process running mem=1020M(199.3%)
2019-04-18T11:23:46.561420+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:24:06.626757+00:00 heroku[worker.1]: Process running mem=1020M(199.3%)
2019-04-18T11:24:06.626922+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:24:27.698419+00:00 heroku[worker.1]: Process running mem=1020M(199.3%)
2019-04-18T11:24:27.698526+00:00 heroku[worker.1]: Error R14 (Memory quota exceeded)
2019-04-18T11:24:47.311199+00:00 heroku[worker.1]: Process running mem=1279M(249.9%)
```

I'm not sure there's much we can do to easily fix this unless you can truncate that author list (they won't all render in the paper anyway)?
