### 2019-04-18 480e16f
Removed most authors from Planck papers
* Looks like the trouble submitting to JOSS was the long list of
authors on the Planck papers in the .bib.
