+++ paper/paper.bib
@@ -282,87 +282,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Planck13,
-       author = {{Planck Collaboration} and {Ade}, P.~A.~R. and {Aghanim}, N. and
-         {Armitage-Caplan}, C. and {Arnaud}, M. and {Ashdown}, M. and {Atrio-Barand
-        ela}, F. and {Aumont}, J. and {Baccigalupi}, C. and {Banday}, A.~J. and
-         {Barreiro}, R.~B. and {Bartlett}, J.~G. and {Battaner}, E. and
-         {Benabed}, K. and {Beno{\^\i}t}, A. and {Benoit-L{\'e}vy}, A. and
-         {Bernard}, J. -P. and {Bersanelli}, M. and {Bielewicz}, P. and
-         {Bobin}, J. and {Bock}, J.~J. and {Bonaldi}, A. and {Bond}, J.~R. and
-         {Borrill}, J. and {Bouchet}, F.~R. and {Bridges}, M. and {Bucher}, M. and
-         {Burigana}, C. and {Butler}, R.~C. and {Calabrese}, E. and
-         {Cappellini}, B. and {Cardoso}, J. -F. and {Catalano}, A. and
-         {Challinor}, A. and {Chamballu}, A. and {Chary}, R. -R. and {Chen}, X. and
-         {Chiang}, H.~C. and {Chiang}, L. -Y. and {Christensen}, P.~R. and
-         {Church}, S. and {Clements}, D.~L. and {Colombi}, S. and
-         {Colombo}, L.~P.~L. and {Couchot}, F. and {Coulais}, A. and
-         {Crill}, B.~P. and {Curto}, A. and {Cuttaia}, F. and {Danese}, L. and
-         {Davies}, R.~D. and {Davis}, R.~J. and {de Bernardis}, P. and
-         {de Rosa}, A. and {de Zotti}, G. and {Delabrouille}, J. and
-         {Delouis}, J. -M. and {D{\'e}sert}, F. -X. and {Dickinson}, C. and
-         {Diego}, J.~M. and {Dolag}, K. and {Dole}, H. and {Donzelli}, S. and
-         {Dor{\'e}}, O. and {Douspis}, M. and {Dunkley}, J. and {Dupac}, X. and
-         {Efstathiou}, G. and {Elsner}, F. and {En{\ss}lin}, T.~A. and
-         {Eriksen}, H.~K. and {Finelli}, F. and {Forni}, O. and {Frailis}, M. and
-         {Fraisse}, A.~A. and {Franceschi}, E. and {Gaier}, T.~C. and
-         {Galeotta}, S. and {Galli}, S. and {Ganga}, K. and {Giard}, M. and
-         {Giardino}, G. and {Giraud-H{\'e}raud}, Y. and {Gjerl{\o}w}, E. and
-         {Gonz{\'a}lez-Nuevo}, J. and {G{\'o}rski}, K.~M. and {Gratton}, S. and
-         {Gregorio}, A. and {Gruppuso}, A. and {Gudmundsson}, J.~E. and
-         {Haissinski}, J. and {Hamann}, J. and {Hansen}, F.~K. and {Hanson}, D. and
-         {Harrison}, D. and {Henrot-Versill{\'e}}, S. and
-         {Hern{\'a}ndez-Monteagudo}, C. and {Herranz}, D. and {Hildebrand
-        t}, S.~R. and {Hivon}, E. and {Hobson}, M. and {Holmes}, W.~A. and
-         {Hornstrup}, A. and {Hou}, Z. and {Hovest}, W. and
-         {Huffenberger}, K.~M. and {Jaffe}, A.~H. and {Jaffe}, T.~R. and
[…]
