+++ paper/paper.bib
@@ -282,7 +282,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Planck13,
-       author = {Planck Collaboration},
+       author = {{Planck Collaboration}},
         title = "{Planck 2013 results. XVI. Cosmological parameters}",
       journal = {A&A},
      keywords = {cosmic background radiation, cosmological parameters, early Universe, inflation, primordial nucleosynthesis, Astrophysics - Cosmology and Nongalactic Astrophysics},
@@ -298,7 +298,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Planck15,
-       author = {Planck Collaboration},
+       author = {{Planck Collaboration}},
         title = "{Planck 2015 results. XIII. Cosmological parameters}",
       journal = {A&A},
      keywords = {cosmology: observations, cosmology: theory, cosmic background radiation, cosmological parameters, Astrophysics - Cosmology and Nongalactic Astrophysics},
@@ -314,7 +314,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Planck18,
-       author = {Planck Collaboration},
+       author = {{Planck Collaboration}},
         title = "{Planck 2018 results. VI. Cosmological parameters}",
       journal = {arXiv e-prints},
      keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
