### 2019-04-18 fb5e003
Reverting back to no logo and fixing typo
