### 2019-04-18 f7467cb
Reverting adding fruitbat logo
