### 2019-04-18 #1377 @arfon
That's looking better. Feel free to add back in the logo now as I don't think this was the source of the problem.
