### 2019-04-18 246e2e0
Added smaller logo to paper
