### 2019-04-18 #1377 @abatten
Actually it was also an issue. I tried with the logo and it couldn't generate the PDF for 14 minutes, so I have re-scaled the image to see if that now works.
