### 2019-04-18 #1377 @abatten
Alright! Looking good, I'll merge to paper development branch into master.
