### 2019-04-22 #1399 @whedon
[REVIEW]: Fruitbat: A Python Package for Estimating Redshifts of Fast Radio Bursts
**Submitting author:** @abatten (<a href="http://orcid.org/0000-0001-7599-6488">Adam Batten</a>)
**Repository:** <a href="https://github.com/abatten/fruitbat" target ="_blank">https://github.com/abatten/fruitbat</a>
**Version:** v1.0.1
