### 2019-04-22 #1377 @xuanxu
@zingale, please head over to #1399 for the actual review
