+++ docs/source/user_guide/getting_started.rst
@@ -34,8 +34,6 @@ each requirement.
 
  - pandas: Reading csv files from FRBCAT
 
- - pybind11: Python bindings for C++. Needed for pyymw16.
-
  - pyymw16: Python wrapper for YMW16 galactic dispersion measure model.
 
  - e13tools: Utility tools for writing docstrings.
