+++ requirements.txt
@@ -3,6 +3,5 @@ scipy>=1.0.0
 astropy>=2.0.0
 matplotlib>=2.2.4
 pandas>=0.22.0
-pybind11>=2.2
-pyymw16>=2.0.4
+pyymw16>=2.0.5
 e13tools>=0.5.3
