### 2019-04-28 8d176e1
Added bibtex for the JOSS paper
