### 2019-04-28 abatten/fruitbat#27 @zingale
update dependencies
In the `README.md`, please consider adding`pybind11` to the list of dependencies.
