### 2019-04-28 #1399 @zingale
added an issue about installing: https://github.com/abatten/fruitbat/issues/27
