### 2019-04-28 abatten/fruitbat#27 @zingale
also please consider adding this to the sphinx documentation in the "Getting Started" section.
