### 2019-04-28 abatten/fruitbat#29 @zingale
unpickling error when doing the "simple example"
When I run the "simple example" here:

https://fruitbat.readthedocs.io/en/latest/user_guide/getting_started.html#simple-example

I get an `UnpicklingError`:

```
FRB180110.calc_redshift(method="Zhang2018", cosmology="Planck18")
UnpicklingError                           Traceback (most recent call last)
/usr/lib64/python3.7/site-packages/numpy/lib/npyio.py in load(file, mmap_mode, allow_pickle, fix_imports, encoding)
    439             try:
--> 440                 return pickle.load(fid, **pickle_kwargs)
    441             except Exception:

UnpicklingError: invalid load key, 'v'.

OSError                                   Traceback (most recent call last)
<ipython-input-5-9a85f326ef82> in <module>
----> 1 FRB180110.calc_redshift(method="Zhang2018", cosmology="Planck18")

~/.local/lib/python3.7/site-packages/fruitbat/_frb.py in calc_redshift(self, method, cosmology, subtract_host, lookup_table)
    306                 table_name = "".join(["custom_", method, ".npz"])
    307
--> 308             lookup_table = table.load(table_name)
    309
    310         self.z = table.get_z_from_table(input_dm, lookup_table)

~/.local/lib/python3.7/site-packages/fruitbat/table.py in load(name, data_dir)
    138
    139     filename = os.path.join(data_dir, name)
--> 140     return np.load(filename)
    141
    142

/usr/lib64/python3.7/site-packages/numpy/lib/npyio.py in load(file, mmap_mode, allow_pickle, fix_imports, encoding)
    441             except Exception:
    442                 raise IOError(
--> 443                     "Failed to interpret file %s as a pickle" % repr(file))
    444     finally:
    445         if own_fid:

OSError: Failed to interpret file '/home/zingale/.local/lib/python3.7/site-packages/fruitbat/data/Zhang2018_Planck18.npz' as a pickle
```

I am using python 3.7.3
