### 2019-04-28 #1399 @zingale
adding an issue about reproducing the simple example: https://github.com/abatten/fruitbat/issues/29
