### 2019-04-29 abatten/fruitbat#29 @abatten
Hi @zingale, I just tested it with python 3.7.3, and I can't seem to replicate your issue. This might be an issue with the version of `numpy`. Are you able to tell me which version of `numpy` you are using?

Also, did you install `fruitbat` using `pip` or by cloning the master branch?
