### 2019-04-29 abatten/fruitbat#29 @zingale
I'm running numpy 1.15.4

playing around some more, things work if I install the package via:
```
pip3 install fruitbat --user
```
but I get the above failures if I install from source, as:
```
git clone https://github.com/abatten/fruitbat
cd fruitbat/
pip3 install . --user
```
so maybe there is a step missing in the latter install method?
