### 2019-04-30 abatten/fruitbat#27 @abatten
In pull request #25 I originally added `pybind11` to the list of requirements to avoid the `Failed building wheel for pyymw16` error even though it correctly install it appears that it didn't. However this didn't seem to resolve the issue as it is an issue with the `pyymw16` package. The author of the pyymw16 package appears to have a fix (telegraphic/pyymw16#1), I am waiting for the changes to be made available on pip.

`fruitbat` should still install correctly without `pybind11` even though it displays an error warning.
