### 2019-04-30 da3f13f
Added link to source code and PyPI in docs
* Added source code link in docs
* Added JOSS badge to README
