+++ README.rst
@@ -1,4 +1,4 @@
-|PyPI| |Python| |License| |Travis| |Docs| |CodeCov|
+|PyPI| |Python| |License| |Travis| |Docs| |CodeCov| |JOSS|
 
 |Logo|
 
@@ -152,7 +152,13 @@ reference our paper.
 
 .. |CodeCov| image:: https://codecov.io/gh/abatten/fruitbat/branch/master/graph/badge.svg
     :target: https://codecov.io/gh/abatten/fruitbat
+    :alt: Code Coverage
 
 .. |License| image:: https://img.shields.io/pypi/l/fruitbat.svg?colorB=purple&label=License
     :target: https://github.com/abatten/fruitbat/raw/master/LICENSE
     :alt: PyPI - License
+
+.. |JOSS| image:: http://joss.theoj.org/papers/634bb69f2445c7457bea5dbc0b83e650/status.svg
+    :target: http://joss.theoj.org/papers/634bb69f2445c7457bea5dbc0b83e650
+    :alt: JOSS Review Status
+
