### 2019-04-30 abatten/fruitbat#29 @abatten
After testing this for a long time I think I have worked out the issue. It appears that you do not have git-lfs installed.

The data files for `fruitbat` are saved in .npz format and are stored using the git Large File Storage. So when you clone the directory and install it, if you don't have git-lfs installed python doesn't know how to extract the data.

I'll add some documentation to the README and the online docs listing this a a requirement if you want to install from source, and links describing the process.

There is a wiki on installing git-lfs depending on your OS. https://github.com/git-lfs/git-lfs/wiki/Installation It's easily install using homebrew (macOS) or apt-get (debian).

You may/may not have to clone the directory again.
