### 2019-05-01 abatten/fruitbat#31 @zingale
comoving distance differs from docs
Working through:

https://fruitbat.readthedocs.io/en/latest/user_guide/using_fruitbat.html

when I get to "Calculating Distances" and the first subsection, "Comoving Distance", I get
```
In [22]: frb.calc_comoving_distance()
Out[22]: <Quantity 1812.21526231 Mpc>
```
instead of the 2020.297 value reported.  Once I create the next `frb` object in the code block below this, if I go back and look at `calc_comoving_distance()` I get the value in the docs.  So perhaps the docs were executed out of order?
