### 2019-05-01 #1399 @zingale
There are two more issues to be resolved:

https://github.com/abatten/fruitbat/issues/32
https://github.com/abatten/fruitbat/issues/31

otherwise, it looks good.
