### 2019-05-02 abatten/fruitbat#31 @abatten
Ahh, I think at one point I wrote luminosity distance before the comoving distance and I rearranged then to be in alphabetical order which causes that mistake. This should be a simple fix, I'll add the same code from luminosity distance to the comoving distance section.
