### 2019-05-02 abatten/fruitbat#32 @abatten
This is a typo in the documentation. It should be `fruitbat.cosmologies`.
