### 2019-05-02 9d86f8e
Added instructions for installing git-lfs
* Added git-lfs instructions
* Cleaned up some white space
* Fixed typo
