### 2019-05-02 cd0174c
Added git-lfs installation instructions
