### 2019-05-02 759f4d7
Fixed incorrect documentation for functions
* Fixed comoving distance to use the correct FRB object.
* Fixed typo in creating custom cosmologies. cosmology -> cosmologies
