### 2019-05-02 feff80f
Close the lookup table to avoid any memory issues.
