### 2019-05-02 abatten/fruitbat#32 @abatten
This has been fixed in #33.
