### 2019-05-02 abatten/fruitbat#31 @abatten
This has been fixed in #33.
