### 2019-05-02 abatten/fruitbat#29 @abatten
The documentation for installing git-lfs has been added in #33.
