### 2019-05-02 #1399 @abatten
I have fixed all the issues that you mentioned it the last couple of pull requests.
