### 2019-05-02 #1399 @abatten
Whilst we are here, do you know if it is possible to get the TeX files that produce the JOSS paper? This would be required if I want to submit the paper to the ArXiv. I just tested it with the PDF created by Whedon and it was rejected, requesting for all the TeX files instead of a PDF.
