### 2019-05-02 #1399 @kyleniemeyer
@abatten regarding the TeX files, unfortunately the conversion using Pandoc happens automatically with our whedon bot, so it isn't *super* easy to get those files. You may be able to do the conversion yourself manually based on the LaTeX template here: https://github.com/openjournals/whedon/tree/master/resources
