### 2019-05-02 #1399 @xuanxu
@abatten Could you please release a new version with all the recent changes?
Once you do that I need you to create an archive of that version in zenodo/figshare/other and report the DOI here.
