+++ paper/paper.bib
@@ -145,7 +145,7 @@ archivePrefix = {arXiv},
          {Ransom}, S.~M. and {Scholz}, P. and {Seymour}, A. and
          {Spitler}, L.~G. and {van Langevelde}, H.~J. and {Wharton}, R.~S.},
         title = "{The Host Galaxy and Redshift of the Repeating Fast Radio Burst FRB 121102}",
-      journal = {\apj},
+      journal = {ApJ},
      keywords = {galaxies: distances and redshifts, galaxies: dwarf, galaxies: ISM, stars: magnetars, stars: neutron, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = "2017",
         month = "Jan",
