+++ requirements.txt
@@ -3,5 +3,5 @@ scipy>=1.0.0
 astropy>=2.0.0
 matplotlib>=2.2.4
 pandas>=0.22.0
-pyymw16>=2.0.5
+pyymw16>=2.0.7
 e13tools>=0.5.3
