### 2019-05-05 #1399 @abatten
I made a new release: 1.0.1 so this will have to be updated in the paper.
Archive version: 10.5281/zenodo.2667568
