### 2019-05-05 #1399 @xuanxu
@abatten I can't find the release listed in GitHub: https://github.com/abatten/fruitbat/releases
