### 2019-05-05 #1399 @abatten
Ah I fixed that. There should be the new release 1.0.1.

Also, I messed up the files in the Zenodo archive so I had to make a new zenodo doi: 10.5281/zenodo.2667596
