### 2019-05-05 #1399 @xuanxu
This is ready now for publication.
Pinging EIC for final acceptance: @openjournals/joss-eics
