### 2019-05-05 #1399 @kyleniemeyer
Thanks @xuanxu. CrossRef is having some issues right now, so we are going to pause on publishing accepted articles until those issues are resolved.
