### 2019-05-09 #1399 @labarba
Congratulations, @abatten, your JOSS paper is now published! 🎉

With huge thanks to our editor: @xuanxu, and the reviewer: @zingale — we couldn't do this without you! 🙏
